refactor: splited reducer into slices
