+++ src/app/consts.ts
@@ -1,7 +1,7 @@
 export enum AuthorizationStatus {
-  Auth = 'AUTH',
-  NoAuth = 'NO_AUTH',
-  Unknown = 'UNKNOWN',
+  AUTH = 'AUTH',
+  NO_AUTH = 'NO_AUTH',
+  UNKNOWN = 'UNKNOWN',
 }
 
 export const enum FetchStatus {
@@ -9,4 +9,11 @@ export const enum FetchStatus {
   LOADING = 'LOADING',
   SUCCESS = 'SUCCESS',
   FAILURE = 'FAILURE',
+}
+
+export const enum NameSpace {
+  OFFER = 'OFFER',
+  USER = 'USER',
+  CITY = 'CITY',
+  REVIEW = 'REVIEW',
 }
