+++ src/app/logged-route/index.tsx
@@ -1,13 +1,11 @@
 import { Navigate } from 'react-router-dom';
 
+import { getIsAuthenticated } from '../../entities/user';
 import { WithChildren } from '../../shared/interfaces';
-import { AuthorizationStatus } from '../consts';
 import { useAppSelector } from '../store/hooks';
 
 export const LoggedRoute: React.FC<WithChildren> = ({ children }) => {
-  const isAuthorizated = useAppSelector(
-    (state) => state.authorizationStatus === AuthorizationStatus.Auth,
-  );
+  const isAuthorizated = useAppSelector(getIsAuthenticated);
 
   return isAuthorizated ? children : <Navigate to={'/login'} />;
 };
