+++ src/app/store/interfaces.ts
@@ -1,26 +1,13 @@
-import { City } from '../../entities/city';
-import { OfferMaximum, OfferPreview } from '../../entities/offer-card';
-import { CommentGet } from '../../entities/review';
-import { User } from '../../entities/user';
-import { AuthorizationStatus, FetchStatus } from '../consts';
+import { AxiosInstance } from 'axios';
+
 import { store } from '.';
 
-export type State = {
-  city: City;
-
-  offers?: OfferPreview[];
-  offersFetchStatus: FetchStatus;
-
-  user?: User;
-  authorizationStatus: AuthorizationStatus;
-
-  offer?: OfferMaximum;
-  offerFetchStatus: FetchStatus;
-
-  reviews?: CommentGet[];
-  reviewsFetchStatus: FetchStatus;
-
-  activeOfferId?: OfferPreview['id'];
-};
+export type State = ReturnType<typeof store.getState>;
 
 export type AppDispatch = typeof store.dispatch;
+
+export type DispatchStateExtra = {
+  dispatch: AppDispatch;
+  getState: () => State;
+  extra: AxiosInstance;
+};
