+++ src/entities/city/index.ts
@@ -2,3 +2,6 @@ export type { CityName, City } from './interfaces';
 export { CityNames } from './interfaces';
 export { cities } from './const';
 export { CitiesList } from './cities-list';
+
+export * from './model/selectors';
+export * from './model/reducer';
