+++ src/entities/city/model/selectors.ts
@@ -1,0 +1,4 @@
+import { NameSpace } from '../../../app/consts';
+import { State } from '../../../app/store/interfaces';
+
+export const getCity = (state: State) => state[NameSpace.CITY].city;
