+++ src/entities/offer-card/index.tsx
@@ -2,3 +2,10 @@ export type { OfferPreview, OfferMaximum } from './interfaces';
 export { OfferCard } from './ui/offer-card';
 export { OffersByCities } from './ui/offers-by-cities';
 export { OffersList } from './ui/offers-list';
+
+// eslint-disable-next-line react-refresh/only-export-components
+export * from './model/selectors';
+// eslint-disable-next-line react-refresh/only-export-components
+export * from './model/reducer';
+// eslint-disable-next-line react-refresh/only-export-components
+export * from './api/api-actions';
