+++ src/entities/review/index.ts
@@ -1,3 +1,6 @@
 export { Review } from './ui/review';
 export { ReviewsList } from './ui/reviews-list';
 export type { CommentGet } from './interfaces';
+export * from './model/selectors';
+export * from './model/reducer';
+export * from './api/api-actions';
