+++ src/entities/user/index.ts
@@ -1,1 +1,4 @@
 export type { User, UserDto } from './interfaces';
+export * from './model/selectors';
+export * from './model/reducer';
+export * from './api/api-actions';
